Move data classes to entity folder
