+++ moxy-compiler/src/main/java/moxy/compiler/viewstate/entity/StrategyWithTag.kt
@@ -1,4 +1,4 @@
-package moxy.compiler.viewstate
+package moxy.compiler.viewstate.entity
 
 import javax.lang.model.element.TypeElement
