+++ moxy-compiler/src/main/java/moxy/compiler/viewstate/entity/ViewInterfaceMethod.kt
@@ -1,8 +1,9 @@
-package moxy.compiler.viewstate
+package moxy.compiler.viewstate.entity
 
 import com.squareup.javapoet.ParameterSpec
 import moxy.compiler.asTypeElement
 import moxy.compiler.equalsByType
+import moxy.compiler.viewstate.ViewMethodParameters
 import javax.lang.model.element.ExecutableElement
 import javax.lang.model.type.DeclaredType
